+++ src/utils/tree.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 use std::{
-    collections::HashSet,
+    collections::{HashSet, VecDeque},
     fmt::{Debug, Display},
 };
 
@@ -8,14 +8,23 @@ use thiserror::Error;
 
 use crate::{installer::types::PackageId, register::package_register::PackageRegister, repositories::error::RepositoryError};
 
+// Static string prefixes for the tree display
+const BRANCH: &str = "\u{251C}\u{2500}\u{2500}\u{2500} ";
+const LAST_BRANCH: &str = "\u{2514}\u{2500}\u{2500}\u{2500} ";
+const VERTICAL_LINE: &str = "\u{2502}    ";
+const EMPTY_SPACE: &str = "     ";
+
 /// The errors that occur while doing tree operations.
 #[derive(Error, Debug)]
 pub enum TreeError {
     #[error("Package id '{0}' cannot be found")]
     NotFound(PackageId),
 
-    #[error("Cannot create tree, because necessary build attribute '{0}' was missing from the tree builder.")]
-    MissingBuildAttributes(String),
+    #[error("Parent with node id '{0}' cannot be found")]
+    NonExistentParent(usize),
+
+    #[error("'{0}' has itself as a dependency, creating a cycle in the tree")]
+    CycleError(PackageId),
 
     #[error("Cannot create tree, because of an error reading the repository.")]
     RepositoryError(#[from] RepositoryError),
@@ -23,132 +32,110 @@ pub enum TreeError {
 
 pub type Result<T> = std::result::Result<T, TreeError>;
 
+pub struct Tree<V, L: Eq> {
+    // A list of all nodes in the tree. Note that this list will always have parents before children by
+    // way of construction (you can only add a node as child of another node).
+    nodes: Vec<Node<V, L>>,
+}
+
 /// Represents a node in a dependency tree with generic values and labels.
 #[derive(Debug)]
 pub struct Node<V, L: Eq> {
-    id: PackageId,
+    package_id: PackageId,
     value: V,
-    children: Vec<Node<V, L>>,
+    children: Vec<usize>,
+    parent_index: usize,
     label: L,
 }
 
-/// A tree builder to build trees from a root, an expander and a populator.
-#[derive(Debug)]
-pub struct TreeBuilder<E, P, T, V, L: Eq>
-where
-    E: Fn(&Node<V, L>) -> Result<Vec<T>>,
-    P: Fn(T) -> Result<(PackageId, V, L)>,
-{
-    /// The root of the tree.
-    root: Option<Node<V, L>>,
+impl<V, L: Eq> Tree<V, L> {
+    pub fn new(root: Node<V, L>) -> Self {
+        Self { nodes: vec![root] }
+    }
 
-    /// Closure specifying how to get children of a certain package id.
-    expander: Option<E>,
-
-    /// Closure specifying how to populate the tree with the expander return values.
-    populator: Option<P>,
-}
-
-// Static string prefixes for the tree display
-const BRANCH: &str = "\u{251C}\u{2500}\u{2500}\u{2500} ";
-const LAST_BRANCH: &str = "\u{2514}\u{2500}\u{2500}\u{2500} ";
-const VERTICAL_LINE: &str = "\u{2502}    ";
-const EMPTY_SPACE: &str = "     ";
-
-impl<E, P, T, V, L: Eq> TreeBuilder<E, P, T, V, L>
-where
-    E: Fn(&Node<V, L>) -> Result<Vec<T>>,
-    P: Fn(T) -> Result<(PackageId, V, L)>,
-{
-    pub fn new() -> Self {
-        Self {
-            root: None,
-            expander: None,
-            populator: None,
+    /// Adds a node to the tree under the given parent index.
+    /// Returns the index of the node added to the nodes list.
+    /// Returns a `TreeError::CycleError` if the new node creates a cycle in the tree.
+    /// Returns a `TreeError::NonExistentParent` if the given parent index doesn't exist.
+    pub fn add_node(&mut self, parent_index: usize, mut node: Node<V, L>) -> Result<usize> {
+        // Check for cycles before adding the node to the tree
+        if self.is_cyclic(&parent_index, node.get_package_id())? {
+            return Err(TreeError::CycleError(node.get_package_id().clone()));
         }
-    }
 
-    pub fn expander(mut self, expander: E) -> Self {
-        self.expander = Some(expander);
-        self
-    }
-
-    pub fn populator(mut self, populator: P) -> Self {
-        self.populator = Some(populator);
-        self
-    }
-
-    /// Sets the root of the tree. Its children are initialized to an empty vec.
-    pub fn root(mut self, package_id: PackageId, value: V, label: L) -> Self {
-        self.root = Some(Node {
-            id: package_id,
-            value,
-            children: Vec::new(),
-            label,
-        });
-
-        self
-    }
-
-    /// Builds the tree. Returns a result containing the root if successful. If any of
-    /// the tree builder attributes are None a MissingBuildAttributes error is returned instead.
-    pub fn build(self) -> Result<Node<V, L>> {
-        let mut root = match self.root {
-            Some(root) => root,
-            None => return Err(TreeError::MissingBuildAttributes("root".to_string())),
+        // Get the parent and check if it exists
+        let parent = match self.nodes.get_mut(parent_index) {
+            Some(parent) => parent,
+            None => return Err(TreeError::NonExistentParent(parent_index)),
         };
 
-        let expander = match &self.expander {
-            Some(expander) => expander,
-            None => return Err(TreeError::MissingBuildAttributes("expander".to_string())),
-        };
+        node.parent_index = parent_index;
+        self.nodes.push(node);
+        let index = self.nodes.len() - 1;
 
-        let populator = match &self.populator {
-            Some(populator) => populator,
-            None => return Err(TreeError::MissingBuildAttributes("populator".to_string())),
-        };
-
-        root.expand(&expander, &populator)?;
-        Ok(root)
+        parent.children.push(index);
+        Ok(index)
     }
-}
 
-// Generic node implementation.
-impl<V, L: Eq> Node<V, L> {
-    /// Expands a node. The expander is used in combination with the populator to get its child nodes.
-    /// If a child already exists it's not added again.
-    pub fn expand<E, P, T>(&mut self, expander: &E, populator: &P) -> Result<()>
+    /// Gets the root of the tree.
+    #[expect(unused)]
+    pub fn get_root(&self) -> &Node<V, L> {
+        self.nodes.get(0).expect("Expected root to exist")
+    }
+
+    /// Returns a reference to a list which contains the nodes of the tree.
+    pub fn get_nodes(&self) -> &Vec<Node<V, L>> {
+        &self.nodes
+    }
+
+    /// Gets a node based on its index.
+    pub fn get_node_by_index(&self, index: usize) -> Option<&Node<V, L>> {
+        self.nodes.get(index)
+    }
+
+    /// Gets a node mutably based on its index.
+    pub fn get_node_by_index_mut(&mut self, index: usize) -> Option<&mut Node<V, L>> {
+        self.nodes.get_mut(index)
+    }
+
+    /// Gets the ids of the children of the node in a hashset.
+    /// Checks for a filter on the label and will only return the children that satisfy the filter.
+    pub fn get_children_ids_filtered<F>(&self, node: &Node<V, L>, filter: F) -> HashSet<PackageId>
     where
-        E: Fn(&Node<V, L>) -> Result<Vec<T>>,
-        P: Fn(T) -> Result<(PackageId, V, L)>,
+        F: Fn(&L) -> bool,
     {
-        let existing_childs = self.get_children_ids();
+        node.get_children()
+            .iter()
+            .map(|c| self.get_node_by_index(c).expect("Expected node to exist"))
+            .filter(|c| filter(&c.label))
+            .map(|c| c.package_id.clone())
+            .collect()
+    }
 
-        for child in expander(self)? {
-            let (id, value, label) = populator(child)?;
-
-            // Node already exists, skip adding
-            if existing_childs.contains(&id) {
-                continue;
-            }
-
-            let mut child_node = Node {
-                id,
-                value,
-                children: Vec::new(),
-                label,
+    /// Checks if a given package id will create a cycle in the tree.
+    /// Returns true if a cycle will be formed, false if not.
+    /// Returns `TreeError::NonExistentParent` if the given parent_index doesn't exist.
+    fn is_cyclic(&self, parent_index: &usize, package_id: &PackageId) -> Result<bool> {
+        let mut current_parent = parent_index;
+        while *current_parent != 0 {
+            let Some(parent_node) = self.get_node_by_index(current_parent) else {
+                return Err(TreeError::NonExistentParent(*current_parent));
             };
 
-            child_node.expand(expander, populator)?;
-            self.children.push(child_node);
+            if parent_node.get_package_id() == package_id {
+                return Ok(true);
+            }
+
+            current_parent = parent_node.get_parent();
         }
 
-        Ok(())
+        Ok(false)
     }
 
     /// The implementation of the tree display.
-    fn display_impl(&self, f: &mut std::fmt::Formatter<'_>, node: &Node<V, L>, prefix: &str) -> std::fmt::Result {
-        writeln!(f, "{prefix}{}", node.id)?;
+    fn display_impl(&self, f: &mut std::fmt::Formatter<'_>, node_index: usize, prefix: &str) -> std::fmt::Result {
+        let node = self.get_node_by_index(node_index).expect("Expected node to exist");
+        writeln!(f, "{prefix}{}", node.package_id)?;
 
         // Note that when the input prefix is "" then this prefix will also be ""
         let prefix = match prefix.ends_with(BRANCH) {
@@ -167,10 +154,25 @@ impl<V, L: Eq> Node<V, L> {
 
         Ok(())
     }
+}
+
+// Generic node implementation.
+impl<V, L: Eq> Node<V, L> {
+    /// Creates a new `Node`. Note that the parent_index is 0 by default. The root will have itself as parent
+    /// and when adding a node to a `Tree` the parent index will need to be adjusted.
+    pub fn new(package_id: PackageId, value: V, label: L) -> Self {
+        Self {
+            package_id,
+            value,
+            children: Vec::new(),
+            parent_index: 0,
+            label,
+        }
+    }
 
     /// Gets the package id.
-    pub fn get_id(&self) -> &PackageId {
-        &self.id
+    pub fn get_package_id(&self) -> &PackageId {
+        &self.package_id
     }
 
     /// Gets the value of the node.
@@ -178,71 +180,55 @@ impl<V, L: Eq> Node<V, L> {
         &self.value
     }
 
-    /// Gets the ids of the children of the node in a hashset.
-    /// Checks for a filter on the label and will only return the children that satisfy the filter.
-    pub fn get_children_ids_filtered<F>(&self, filter: F) -> HashSet<PackageId>
-    where
-        F: Fn(&L) -> bool,
-    {
-        self.children.iter().filter(|c| filter(&c.label)).map(|c| c.id.clone()).collect()
-    }
-
-    /// Gets the ids of the children of the node in a hashset.
-    pub fn get_children_ids(&self) -> HashSet<PackageId> {
-        self.children.iter().map(|c| c.id.clone()).collect()
+    /// Gets the parent index of the node.
+    pub fn get_parent(&self) -> &usize {
+        &self.parent_index
     }
 
     /// Gets the child nodes as reference.
-    pub fn get_children(&self) -> &Vec<Node<V, L>> {
+    pub fn get_children(&self) -> &Vec<usize> {
         &self.children
-    }
-
-    /// Gets the child nodes as mutable reference.
-    pub fn get_children_mut(&mut self) -> &mut Vec<Node<V, L>> {
-        &mut self.children
     }
 
     /// Gets the label.
     pub fn get_label(&self) -> &L {
         &self.label
     }
-
-    /// Sets a new label.
-    pub fn set_label(&mut self, new_label: L) {
-        self.label = new_label;
-    }
 }
 
-/// Represents an empty node, without any value or label.
-pub type EmptyNode = Node<(), ()>;
+/// Represents a tree, without any value or label at its nodes.
+pub type EmptyTree = Tree<(), ()>;
 
 // An empty node implementation, without any values or labels.
-impl EmptyNode {
+impl EmptyTree {
     /// Builds a simple tree based on the installed packages.
-    pub fn build_simple_tree(package_id: PackageId, register: &PackageRegister) -> Result<EmptyNode> {
-        TreeBuilder::new()
-            .root(package_id, (), ())
-            .expander(|p| EmptyNode::expander(p, register))
-            .populator(EmptyNode::populator)
-            .build()
-    }
+    pub fn new_empty(package_id: PackageId, register: &PackageRegister) -> Result<EmptyTree> {
+        let root = Node::new(package_id, (), ());
+        let mut tree = Tree::new(root);
 
-    /// Gets the children based on the installed packages.
-    fn expander(parent: &EmptyNode, register: &PackageRegister) -> Result<Vec<PackageId>> {
-        let package = register.get_package_version(parent.get_id()).ok_or(TreeError::NotFound(parent.get_id().clone()))?;
-        Ok(package.dependencies.iter().cloned().collect())
-    }
+        let mut package_queue = VecDeque::from([0 as usize]);
+        while let Some(node_index) = package_queue.pop_front() {
+            let node = tree.get_node_by_index_mut(node_index).expect("Expected node to exist");
+            let dependencies = &register
+                .get_package_version(node.get_package_id())
+                .ok_or(TreeError::NotFound(node.get_package_id().clone()))?
+                .dependencies;
 
-    /// Populates nodes with empty values and labels.
-    fn populator(package_id: PackageId) -> Result<(PackageId, (), ())> {
-        Ok((package_id, (), ()))
+            for dependency in dependencies {
+                let new_node = Node::new(dependency.clone(), (), ());
+                let new_index = tree.add_node(node_index, new_node)?;
+                package_queue.push_back(new_index);
+            }
+        }
+
+        Ok(tree)
     }
 }
 
 // Display trait for nice display of a tree.
-impl<V, L: Eq> Display for Node<V, L> {
+impl<V, L: Eq> Display for Tree<V, L> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.display_impl(f, self, "")?;
+        self.display_impl(f, &0, "")?;
         Ok(())
     }
 }